Move gameplay test pieces to own namespace
